campaign-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
redis==4.6.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-grpc==0.42b0
opentelemetry-exporter-jaeger==1.21.0
prometheus-client==0.19.0
structlog==23.2.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
greenlet==3.0.3
grpcio==1.54.0
grpcio-tools==1.54.0